+++ quantify_scheduler/device_under_test/quantum_device.py
@@ -54,19 +54,20 @@ class QuantumDevice(Instrument):
 
         self.elements = ManualParameter(
             "elements",
-            initial_value=list(),
-            vals=validators.Lists(validators.Strings()),
-            docstring="A list containing the names of all elements that"
-            " are located on this QuantumDevice.",
+            initial_value=dict(),
+            vals=validators.Dict(),
+            docstring="A dictionary with the names of all elements located on this"
+            "QuantumDevice as keys and the corresponding `DeviceElement` objects as values.",
             instrument=self,
         )
 
         self.edges = ManualParameter(
             "edges",
-            initial_value=list(),
-            vals=validators.Lists(validators.Strings()),
-            docstring="A list containing the names of all the edges which connect the"
-            " DeviceElements within this QuantumDevice",
+            initial_value=dict(),
+            vals=validators.Dict(),
+            docstring="A dictionary with the names of all the edges which connect the"
+            " `DeviceElement`s within this QuantumDevice as keys and the corresponding"
+            " `Edge` objects as values.",
             instrument=self,
         )
 
@@ -350,8 +351,14 @@ class QuantumDevice(Instrument):
             If key `name` is not present in `self.elements`.
         """
         if name in self.elements():
-            return self.find_instrument(name)
-        raise KeyError(f"'{name}' is not a element of {self.name}.")
+            element = self.elements()[name]
+            if not Instrument.is_valid(element):
+                raise ValueError(
+                    f"{element} of {self.name} is not a valid device element."
+                    f" Please re-create and re-add it."
+                )
+            return element
+        raise KeyError(f"'{name}' is not an element of {self.name}.")
 
     def add_element(
         self,
@@ -373,12 +380,14 @@ class QuantumDevice(Instrument):
             If :code:`element` is not an instance of the base element.
         """
         if element.name in self.elements():
-            raise ValueError(f"'{element.name}' has already been added.")
+            if Instrument.is_valid(self.elements()[element.name]):
+                # Only raise an error if the element is still valid
+                raise ValueError(f"'{element.name}' has already been added.")
 
         if not isinstance(element, DeviceElement):
             raise TypeError(f"{repr(element)} is not a DeviceElement.")
 
-        self.elements().append(element.name)  # list gets updated in place
+        self.elements()[element.name] = element  # dict gets updated in place
 
     def remove_element(self, name: str) -> None:
         """
@@ -390,7 +399,7 @@ class QuantumDevice(Instrument):
             The element name.
         """
 
-        self.elements().remove(name)  # list gets updated in place
+        self.elements().pop(name)  # dict gets updated in place
 
     def get_edge(self, name: str) -> Instrument:
         """
@@ -412,8 +421,14 @@ class QuantumDevice(Instrument):
             If key `name` is not present in `self.edges`.
         """
         if name in self.edges():
-            return self.find_instrument(name)
-        raise KeyError(f"'{name}' is not a edge of {self.name}.")
+            edge = self.edges()[name]
+            if not Instrument.is_valid(edge):
+                raise ValueError(
+                    f"{edge} of {self.name} is not a valid edge."
+                    f" Please re-create and re-add it."
+                )
+            return edge
+        raise KeyError(f"'{name}' is not an edge of {self.name}.")
 
     def add_edge(self, edge: Edge) -> None:
         """
@@ -426,12 +441,14 @@ class QuantumDevice(Instrument):
             'element_0'-'element_1'
         """
         if edge.name in self.edges():
-            raise ValueError(f"'{edge.name}' has already been added.")
+            if Instrument.is_valid(self.edges()[edge.name]):
+                # Only raise an error if the edge is still valid
+                raise ValueError(f"'{edge.name}' has already been added.")
 
         if not isinstance(edge, Edge):
             raise TypeError(f"{repr(edge)} is not a Edge.")
 
-        self.edges().append(edge.name)
+        self.edges()[edge.name] = edge  # dict gets updated in place
 
     def remove_edge(self, edge_name: str) -> None:
         """
@@ -443,4 +460,4 @@ class QuantumDevice(Instrument):
             The edge name.
         """
 
-        self.edges().remove(edge_name)  # list gets updated in place
+        self.edges().pop(edge_name)  # dict gets updated in place